@@ -10,48 +10,89 @@
     <name>Elide Spring Boot Example</name>
     <description>Elide Spring Boot Example</description>
     <groupId>example</groupId>
-    <version>1.0</version>
+    <version>1.0.0</version>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.5.6</version>
+        <version>3.1.3</version>
     </parent>
 
     <properties>
         <jar.finalName>elide-spring-boot</jar.finalName>
-        <postgres.version>42.3.8</postgres.version>
-        <liquibase.version>4.8.0</liquibase.version>
-        <elide.version>6.1.4</elide.version>
-        <spring.version>2.5.6</spring.version>
-        <junit.jupiter.version>5.5.2</junit.jupiter.version>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
+
+        <elide.version>7.0.0</elide.version>
+
+        <springdoc.version>2.2.0</springdoc.version>
+        <logback-access-spring-boot-starter.version>4.0.0</logback-access-spring-boot-starter.version>
+        <snakeyaml.version>2.2</snakeyaml.version>
+
+        <!-- Jetty 11 uses servlet-api 5 and not 6 -->
+        <jakarta.servlet-api.version>5.0.0</jakarta.servlet-api.version>
+
+        <!-- see https://github.com/graalvm/native-build-tools/issues/358 -->
+        <maven-surefire-plugin.version>3.0.0-M5</maven-surefire-plugin.version>
     </properties>
 
-    <dependencies>
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>jakarta.servlet</groupId>
+                <artifactId>jakarta.servlet-api</artifactId>
+                <version>${jakarta.servlet-api.version}</version>
+            </dependency>
 
+            <dependency>
+                <groupId>com.yahoo.elide</groupId>
+                <artifactId>elide-bom</artifactId>
+                <version>${elide.version}</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+
+    <dependencies>
         <!-- Runtime -->
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <version>1.18.10</version>
         </dependency>
+
+        <dependency>
+            <groupId>org.springdoc</groupId>
+            <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+            <version>${springdoc.version}</version>
+       </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-security</artifactId>
-            <version>${spring.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-validation</artifactId>
         </dependency>
 
         <dependency>
             <groupId>com.yahoo.elide</groupId>
             <artifactId>elide-spring-boot-starter</artifactId>
-            <version>${elide.version}</version>
             <!-- Remove Tomcat -->
             <exclusions>
                 <exclusion>
                     <groupId>org.springframework.boot</groupId>
                     <artifactId>spring-boot-starter-tomcat</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>com.vaadin.external.google</groupId>
+                    <artifactId>android-json</artifactId>
+                </exclusion>
+                <!-- Remove commons-logging as it conflicts with spring-jcl -->
+                <exclusion>
+                    <groupId>commons-logging</groupId>
+                    <artifactId>commons-logging</artifactId>
                 </exclusion>
             </exclusions>
         </dependency>
