@@ -3,7 +3,7 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package example.models;
+package example.models.v1;
 
 import com.yahoo.elide.annotation.Include;
 
@@ -13,15 +13,15 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.util.Date;
 
-@Include(rootLevel = false, name = "version", description = "Artifact version.", friendlyName = "Version")
+@Include(rootLevel = false, name = "version", description = "Artifact version.", friendlyName = "VersionV1")
 @Table(name = "artifactversion")
 @Entity
-public class ArtifactVersion {
+public class ArtifactVersionV1 {
     @Id
     private String name = "";
 
     private Date createdAt = new Date();
 
     @ManyToOne
-    private ArtifactProduct artifact;
+    private ArtifactProductV1 artifact;
 }
