@@ -3,37 +3,29 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package example.models;
+
+package example.models.v2;
 
 import com.yahoo.elide.annotation.Include;
 import com.yahoo.elide.graphql.subscriptions.annotations.Subscription;
 import com.yahoo.elide.graphql.subscriptions.annotations.SubscriptionField;
 
-import lombok.Data;
-
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
-import java.util.ArrayList;
-import java.util.List;
+import lombok.Data;
 
-@Include(name = "group", description = "Artifact group.", friendlyName = "Group")
-@Table(name = "artifactgroup")
+@Include(name = "groups", description = "Artifact group.", friendlyName = "Group")
 @Entity
 @Subscription
 @Data
-public class ArtifactGroup {
+@Table(name = "artifactgroup")
+public class ArtifactGroupV2 {
     @Id
     private String name = "";
 
     @SubscriptionField
-    private String commonName = "";
-
-    @SubscriptionField
-    private String description = "";
-
-    @SubscriptionField
-    @OneToMany(mappedBy = "group")
-    private List<ArtifactProduct> products = new ArrayList<>();
+    @Column(name = "commonName")
+    private String title = "";
 }
